The Theater Seating lab, lab #7
I set it up so if the user ran the program with no arguments it provided the requisite general functionality.  However, if the user passes the work "configure" as an argument to the program then the user has gets an opportunity to change ( add, delete, or edit ) the sections configuration before the general functionality begins.